Add files via upload
In this assignment, we  explored the dataset, which consists 3 different classes. The objective is to build a model for image classification and compare the performance of Manhattan (L1) and Euclidean (L2) distances using 5-fold cross-validation. We  used only the training data, which consists of 3,000 images, and converted  them to grayscale (32x32) to reduce computation time. We then plotted a graph to visualize the accuracy of different hyperparameter values (K) for both distance calculation techniques and discuss which distance calculation technique is better suited for this particular dataset. Furthermore, we  displayed the top 5 predictions made by the model.